Add default matcher for mdc.
* DSL has been removed as does not add any functionality.
