+++ .editorconfig
@@ -14,6 +14,7 @@ indent_style = space
 indent_size = 2
 continuation_indent_size = 4
 wildcard_import_limit = 9999
+curly_bracket_next_line = false
 
 [Makefile]
 indent_style = tab
