+++ build.gradle
@@ -14,7 +14,7 @@ plugins {
 }
 
 group 'com.jspcore'
-version '0.4.7-SNAPSHOT'
+version '0.5.0-SNAPSHOT'
 description = 'A testing library for assert logging messages.'
 
 apply plugin: 'idea'
@@ -31,19 +31,19 @@ repositories {
 }
 
 ext {
-  junit5_version = "5.2.0"
+  junit5_version = "5.5.0"
 }
 
 dependencies {
   api "org.hamcrest:hamcrest-library:2.1"
   implementation "org.slf4j:slf4j-api:1.7.22"
-  implementation "ch.qos.logback:logback-classic:1.1.8"
+  implementation "ch.qos.logback:logback-classic:1.2.3"
   implementation "org.awaitility:awaitility:3.1.6"
   implementation "junit:junit:4.12"
   implementation "org.junit.jupiter:junit-jupiter-api:$junit5_version"
   implementation 'org.spockframework:spock-core:1.3-groovy-2.5'
 
-  testImplementation "org.assertj:assertj-core:3.6.1"
+  testImplementation "org.assertj:assertj-core:3.12.2"
   testImplementation "org.junit.jupiter:junit-jupiter-engine:$junit5_version"
   testImplementation "org.junit.jupiter:junit-jupiter-params:$junit5_version"
   testImplementation "org.junit.vintage:junit-vintage-engine:$junit5_version"
@@ -100,17 +100,13 @@ task sourceJar(type: Jar) {
   from sourceSets.main.allJava
 }
 
-wrapper {
-  gradleVersion = '5.4'
-}
-
 publishing {
   publications {
     MyPublication(MavenPublication) {
       from components.java
       groupId 'com.jspcore'
       artifactId 'logcapture'
-      version '0.4.7'
+      version '0.5.0'
 
       artifacts {
         artifact sourceJar {
@@ -144,8 +140,8 @@ bintray {
     licenses = ['MIT']
     vcsUrl = 'https://github.com/mustaine/logcapture.git'
     version {
-      name = '0.4.7'
-      vcsTag = '0.4.7'
+      name = '0.5.0'
+      vcsTag = '0.5.0'
       released = new Date()
     }
   }
