+++ src/main/java/com/logcapture/assertion/VerificationException.java
@@ -14,18 +14,18 @@ public class VerificationException extends AssertionError {
 
   public static VerificationException forUnmatchedLog(Matcher<List<ILoggingEvent>> expectedLogMessage, List<ILoggingEvent> logEvents) {
     return new VerificationException(String.format(
-      "Expected matching: \n%s\nLogs received: \n%s",
-      expectedLogMessage.toString(),
-      logEvents.stream()
-        .map(VerificationException::formatLogEvent)
-        .collect(Collectors.joining("\n"))
+        "Expected matching: \n%s\nLogs received: \n%s",
+        expectedLogMessage.toString(),
+        logEvents.stream()
+            .map(VerificationException::formatLogEvent)
+            .collect(Collectors.joining("\n"))
     ));
   }
 
   private static String formatLogEvent(ILoggingEvent log) {
     return String.format("level: %s marker: %s mdc: %s message: %s", log.getLevel(),
-                         log.getMarker(),
-                         log.getMDCPropertyMap(),
-                         log.getFormattedMessage());
+        log.getMarker(),
+        log.getMDCPropertyMap(),
+        log.getFormattedMessage());
   }
 }
