+++ src/main/java/com/logcapture/junit/LogCaptureRule.java
@@ -53,7 +53,7 @@ public class LogCaptureRule implements MethodRule, TestRule {
   }
 
   public LogCaptureRule logged(Matcher<List<ILoggingEvent>> expectedLoggingMessage) {
-    new LogCapture<>(logAppender.events(), null).logged(expectedLoggingMessage);
+    new LogCapture<>(logAppender.events()).logged(expectedLoggingMessage);
     return this;
   }
 }
