+++ src/main/java/com/logcapture/matcher/TypedAnythingMatcher.java
@@ -3,7 +3,7 @@ package com.logcapture.matcher;
 import org.hamcrest.BaseMatcher;
 import org.hamcrest.Description;
 
-  public class TypedAnythingMatcher<T> extends BaseMatcher<T> {
+public class TypedAnythingMatcher<T> extends BaseMatcher<T> {
 
   @Override
   public boolean matches(Object item) {
