+++ src/test/java/com/logcapture/assertion/ExpectedLoggedExceptionShould.java
@@ -17,7 +17,7 @@ public class ExpectedLoggedExceptionShould {
   public void match_when_exception_class_match() {
     LoggingEvent logEvent = aLoggingEventWith(INFO, "message", new IllegalArgumentException());
     ExpectedLoggedException expectedLoggedException = ExpectedLoggedException.logException()
-      .withException(isA(IllegalArgumentException.class));
+        .withException(isA(IllegalArgumentException.class));
 
     boolean matches = expectedLoggedException.matches(logEvent);
 
@@ -28,7 +28,7 @@ public class ExpectedLoggedExceptionShould {
   public void match_when_exception_message_match() {
     LoggingEvent logEvent = aLoggingEventWith(INFO, "message", new IllegalArgumentException("message error"));
     ExpectedLoggedException expectedLoggedException = ExpectedLoggedException.logException()
-      .withMessage(equalTo("message error"));
+        .withMessage(equalTo("message error"));
 
     boolean matches = expectedLoggedException.matches(logEvent);
 
@@ -39,7 +39,7 @@ public class ExpectedLoggedExceptionShould {
   public void not_match_when_exception_class_different() {
     LoggingEvent logEvent = aLoggingEventWith(INFO, "message", new IllegalArgumentException());
     ExpectedLoggedException expectedLoggedException = ExpectedLoggedException.logException()
-      .withException(isA(IllegalStateException.class));
+        .withException(isA(IllegalStateException.class));
 
     boolean matches = expectedLoggedException.matches(logEvent);
 
@@ -50,7 +50,7 @@ public class ExpectedLoggedExceptionShould {
   public void not_match_when_exception_message_different() {
     LoggingEvent logEvent = aLoggingEventWith(INFO, "message", new IllegalArgumentException("some error"));
     ExpectedLoggedException expectedLoggedException = ExpectedLoggedException.logException()
-      .withMessage(equalTo("another error"));
+        .withMessage(equalTo("another error"));
 
     boolean matches = expectedLoggedException.matches(logEvent);
 
@@ -60,8 +60,8 @@ public class ExpectedLoggedExceptionShould {
   @Test
   public void to_string_method_when_exception_not_match() {
     ExpectedLoggedException expectedLoggedException = ExpectedLoggedException.logException()
-      .withMessage(equalTo("another error"))
-      .withException(isA(IllegalArgumentException.class));
+        .withMessage(equalTo("another error"))
+        .withException(isA(IllegalArgumentException.class));
 
     assertThat(expectedLoggedException.toString()).isEqualTo("ExpectedLoggedException{expectedMessageMatcher=\"another error\", expectedException=is an instance of java.lang.IllegalArgumentException}");
   }
