+++ src/test/java/com/logcapture/assertion/ExpectedLoggingMessageShould.java
@@ -121,8 +121,8 @@ public class ExpectedLoggingMessageShould {
     LoggingEvent logEvent = aLoggingEventWith(INFO, "message has another message");
 
     ExpectedLoggingMessage expectedLoggingMessage = aLog().withMessage(containsString("message"),
-      containsString("has"),
-      containsString("another")
+        containsString("has"),
+        containsString("another")
     );
 
     boolean matches = expectedLoggingMessage.matches(singletonList(logEvent));
@@ -135,9 +135,9 @@ public class ExpectedLoggingMessageShould {
     LoggingEvent logEvent = aLoggingEventWith(INFO, "message has another message");
 
     ExpectedLoggingMessage expectedLoggingMessage = aLog().withMessage(containsString("message"),
-      containsString("has"),
-      containsString("NO_MATCH"),
-      containsString("another")
+        containsString("has"),
+        containsString("NO_MATCH"),
+        containsString("another")
     );
 
     boolean matches = expectedLoggingMessage.matches(singletonList(logEvent));
@@ -149,7 +149,7 @@ public class ExpectedLoggingMessageShould {
   public void match_when_length_message_match() {
     LoggingEvent logEvent = aLoggingEventWith(INFO, "message");
     ExpectedLoggingMessage expectedLoggingMessage = aLog()
-      .length(equalTo(7));
+        .length(equalTo(7));
 
     boolean matches = expectedLoggingMessage.matches(singletonList(logEvent));
 
@@ -161,7 +161,7 @@ public class ExpectedLoggingMessageShould {
     MDC.put("aKey", "someValue");
     LoggingEvent logEvent = aLoggingEventWith(INFO, "message");
     ExpectedLoggingMessage expectedLoggingMessage = aLog()
-      .withMdc("aKey", equalTo("someValue"));
+        .withMdc("aKey", equalTo("someValue"));
 
     boolean matches = expectedLoggingMessage.matches(singletonList(logEvent));
 
@@ -174,8 +174,8 @@ public class ExpectedLoggingMessageShould {
     MDC.put("anotherKey", "anotherValue");
     LoggingEvent logEvent = aLoggingEventWith(INFO, "message");
     ExpectedLoggingMessage expectedLoggingMessage = aLog()
-      .withMdc("aKey", equalTo("someValue"))
-      .withMdc("anotherKey", equalTo("anotherValue"));
+        .withMdc("aKey", equalTo("someValue"))
+        .withMdc("anotherKey", equalTo("anotherValue"));
 
     boolean matches = expectedLoggingMessage.matches(singletonList(logEvent));
 
@@ -186,7 +186,7 @@ public class ExpectedLoggingMessageShould {
   public void match_when_logger_class_match() {
     LoggingEvent logEvent = aLoggingEventWith(INFO, "message");
     ExpectedLoggingMessage expectedLoggingMessage = aLog()
-      .withLoggerName(equalTo(ExpectedLoggingMessageShould.class.getName()));
+        .withLoggerName(equalTo(ExpectedLoggingMessageShould.class.getName()));
 
     boolean matches = expectedLoggingMessage.matches(singletonList(logEvent));
 
@@ -197,8 +197,8 @@ public class ExpectedLoggingMessageShould {
   public void match_when_exception_class_match() {
     LoggingEvent logEvent = aLoggingEventWith(INFO, "message", new RuntimeException());
     ExpectedLoggingMessage expectedLoggingMessage = aLog()
-      .havingException(logException()
-        .withException(instanceOf(RuntimeException.class)));
+        .havingException(logException()
+            .withException(instanceOf(RuntimeException.class)));
 
     boolean matches = expectedLoggingMessage.matches(singletonList(logEvent));
 
@@ -209,8 +209,8 @@ public class ExpectedLoggingMessageShould {
   public void match_when_message_and_log_level_match() {
     LoggingEvent logEvent = aLoggingEventWith(INFO, "message");
     ExpectedLoggingMessage expectedLoggingMessage = aLog()
-      .withLevel(equalTo(INFO))
-      .withMessage(equalTo("message"));
+        .withLevel(equalTo(INFO))
+        .withMessage(equalTo("message"));
 
     boolean matches = expectedLoggingMessage.matches(singletonList(logEvent));
 
@@ -221,7 +221,7 @@ public class ExpectedLoggingMessageShould {
   public void not_match_when_log_level_different() {
     LoggingEvent logEvent = aLoggingEventWith(INFO, "message");
     ExpectedLoggingMessage expectedLoggingMessage = aLog()
-      .withMessage(equalTo("differentMessage"));
+        .withMessage(equalTo("differentMessage"));
 
     boolean matches = expectedLoggingMessage.matches(singletonList(logEvent));
 
@@ -232,7 +232,7 @@ public class ExpectedLoggingMessageShould {
   public void not_match_when_message_different() {
     LoggingEvent logEvent = aLoggingEventWith(INFO, "message");
     ExpectedLoggingMessage expectedLoggingMessage = aLog()
-      .withMessage(equalTo("anotherMessage"));
+        .withMessage(equalTo("anotherMessage"));
 
     boolean matches = expectedLoggingMessage.matches(singletonList(logEvent));
 
@@ -243,7 +243,7 @@ public class ExpectedLoggingMessageShould {
   public void not_match_when_message_length_different() {
     LoggingEvent logEvent = aLoggingEventWith(INFO, "message");
     ExpectedLoggingMessage expectedLoggingMessage = aLog()
-      .length(equalTo(8));
+        .length(equalTo(8));
 
     boolean matches = expectedLoggingMessage.matches(singletonList(logEvent));
 
@@ -255,7 +255,7 @@ public class ExpectedLoggingMessageShould {
     MDC.put("aKey", "differentValue");
     LoggingEvent logEvent = aLoggingEventWith(INFO, "message");
     ExpectedLoggingMessage expectedLoggingMessage = aLog()
-      .withMdc("aKey", equalTo("someValue"));
+        .withMdc("aKey", equalTo("someValue"));
 
     boolean matches = expectedLoggingMessage.matches(singletonList(logEvent));
 
@@ -269,8 +269,8 @@ public class ExpectedLoggingMessageShould {
     MDC.put("anotherKey", "anotherValue");
     LoggingEvent logEvent = aLoggingEventWith(INFO, "message");
     ExpectedLoggingMessage expectedLoggingMessage = aLog()
-      .withMdc("aKey", equalTo("unmatchedValue"))
-      .withMdc("anotherKey", equalTo("anotherValue"));
+        .withMdc("aKey", equalTo("unmatchedValue"))
+        .withMdc("anotherKey", equalTo("anotherValue"));
 
     boolean matches = expectedLoggingMessage.matches(singletonList(logEvent));
 
@@ -282,7 +282,7 @@ public class ExpectedLoggingMessageShould {
   public void not_match_when_logger_class_different() {
     LoggingEvent logEvent = aLoggingEventWith(INFO, "message");
     ExpectedLoggingMessage expectedLoggingMessage = aLog()
-      .withLoggerName(equalTo("anotherClassName"));
+        .withLoggerName(equalTo("anotherClassName"));
 
     boolean matches = expectedLoggingMessage.matches(singletonList(logEvent));
 
@@ -295,7 +295,7 @@ public class ExpectedLoggingMessageShould {
     logEvent.setMarker(MarkerFactory.getMarker("A_MARKER"));
 
     ExpectedLoggingMessage expectedLoggingMessage = aLog()
-      .withMarker(MarkerFactory.getMarker("A_MARKER"));
+        .withMarker(MarkerFactory.getMarker("A_MARKER"));
 
     boolean matches = expectedLoggingMessage.matches(singletonList(logEvent));
 
@@ -308,7 +308,7 @@ public class ExpectedLoggingMessageShould {
     logEvent.setMarker(MarkerFactory.getMarker("A_MARKER"));
 
     ExpectedLoggingMessage expectedLoggingMessage = aLog()
-      .withMarker("A_MARKER");
+        .withMarker("A_MARKER");
 
     boolean matches = expectedLoggingMessage.matches(singletonList(logEvent));
 
@@ -321,7 +321,7 @@ public class ExpectedLoggingMessageShould {
     logEvent.setMarker(MarkerFactory.getMarker("A_MARKER"));
 
     ExpectedLoggingMessage expectedLoggingMessage = aLog()
-      .withMarker("ANOTHER_MARKER");
+        .withMarker("ANOTHER_MARKER");
 
     boolean matches = expectedLoggingMessage.matches(singletonList(logEvent));
 
@@ -331,24 +331,24 @@ public class ExpectedLoggingMessageShould {
   @Test
   public void describe_failure_using_to_string() {
     ExpectedLoggingMessage expectedLoggingMessage = aLog()
-      .withLevel(equalTo(ERROR))
-      .withMessage(equalTo("message"))
-      .length(equalTo(8))
-      .withMdc("aKey", equalTo("some"))
-      .withMarker("A_MARKER")
-      .withLoggerName(equalTo("className"))
-      .havingException(logException()
-        .withMessage(equalTo("exception thrown"))
-        .withException(causeOf(IllegalArgumentException.class)));
+        .withLevel(equalTo(ERROR))
+        .withMessage(equalTo("message"))
+        .length(equalTo(8))
+        .withMdc("aKey", equalTo("some"))
+        .withMarker("A_MARKER")
+        .withLoggerName(equalTo("className"))
+        .havingException(logException()
+            .withMessage(equalTo("exception thrown"))
+            .withException(causeOf(IllegalArgumentException.class)));
 
     assertThat(expectedLoggingMessage.toString())
-      .contains("logLevelMatcher=<ERROR>")
-      .contains("markerMatcher=<A_MARKER>")
-      .contains("expectedMessageMatcher=[\"message\"]")
-      .contains("expectedLengthMatcher=<8>")
-      .contains("expectedMdc={aKey=\"some\"}")
-      .contains("expectedLoggerNameMatcher=\"className\"")
-      .contains("expectedLoggedException=ExpectedLoggedException{expectedMessageMatcher=\"exception thrown\", expectedException=Expecting exception to be instance of class java.lang.IllegalArgumentException");
+        .contains("logLevelMatcher=<ERROR>")
+        .contains("markerMatcher=<A_MARKER>")
+        .contains("expectedMessageMatcher=[\"message\"]")
+        .contains("expectedLengthMatcher=<8>")
+        .contains("expectedMdc={aKey=\"some\"}")
+        .contains("expectedLoggerNameMatcher=\"className\"")
+        .contains("expectedLoggedException=ExpectedLoggedException{expectedMessageMatcher=\"exception thrown\", expectedException=Expecting exception to be instance of class java.lang.IllegalArgumentException");
   }
 
   private LoggingEvent aLoggingEventWith(Level level, String message) {
